html/semantics/embedded-content/the-iframe-element/iframe-loading-lazy-history-replaceState.html

<!DOCTYPE html>
<title>History state change for iframe loading='lazy' before it is loaded: history.replaceState</title>
<iframe data-src="about:blank#foo" src="support/blank.htm?src" loading="lazy"></iframe>
<script>
const iframe = document.querySelector('iframe');
let replaceStateSuccess = true;
try {
  // Should have no effect on lazy-loading.
  // Per https://html.spec.whatwg.org/#can-have-its-url-rewritten
  // only the fragment can be changed for about: URLs.
  iframe.contentWindow.history.replaceState(null, "", iframe.dataset.src);
} catch(ex) {
  replaceStateSuccess = false;
}
const locationAfterReplaceState = iframe.contentWindow.location.href;
</script>
<!-- Loading testharness.js here is intentional to reproduce a bug in WebKit. -->
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script>
setup({single_test: true});
assert_true(replaceStateSuccess);
assert_equals(locationAfterReplaceState, new URL("about:blank#foo", location.href).href);
onload = () => {
  step_timeout(() => {
    assert_equals(iframe.contentWindow.location.href, new URL("support/blank.htm?src", location.href).href);
    done();
  }, 1000);
}
</script>
<!-- Delay the load event for the same URLs as the potential lazy iframe navigations
     to reduce unstable test results. -->
<iframe src="support/blank.htm?src"></iframe>
